added MIP(Mixed Integer Programming model) solver in algorithms.py as a baseline. for small n in [small_ns = [8, 12, 16]], (threshold set as n=18) run both eact solvers. compare the n-bins returned by both my-own-exact-solver and MIP.
The output of both always matches until n=16, when the input type is large items, MIP is taking forever to give a solution. So I exited the program with ctrl+c. Have stored the output result in output_sample_0.txt

Also, replace all the apperances of this word "strip" with "bin", since we are discussing the 1-D bin problem, otherwise it will be misleading and causing confusions.