Reorganized the files.
Added a parse function to parse the experiment output to csv. The reason I didn't directly write the values into a csv during the experiment or draw the visualization graphs right after the experiment is the experiment itself is already taking a very long time, that way will makes the debugging multiple times harder.

Added a folder for presentation, including a function used to create QR code.